requirements.txt
Flask==3.0.0
Pillow==10.1.0
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
